+++ components/about.js
@@ -2,10 +2,10 @@ import Skills from './sckills'
 
 export default function About (){
     return (
-        <div className='bg-[#161922]  py-40 px-60'>
+        <div className='bg-[#161922] py-40 px-60 relative'>
             <div className="flex">
                 <div className="w-2/4">
-                    <h1 className="text-[9rem]">3<span className="text-[#ffd175]">.</span></h1>
+                    <h1 className="text-[11rem] font-bold">3<span className="text-[#ffd175]">.</span></h1>
                     <div>
                         <h2>Years</h2>
                         <h2>Experience</h2> 
@@ -19,7 +19,7 @@ export default function About (){
                         digging into hard problems, learning new languages and 
                         technologies.</p>
                     <h1 className="name text-[#ffd175] text-5xl my-4">joel kalema</h1>
-                    <button type='button' className="bg-[#ffd175] my-2">Hire me</button>
+                    <button type='button' className="bg-[#ffd175] my-10">Hire me</button>
                 </div>
             </div>
             <Skills />
