+++ components/portfolio.js
@@ -1,0 +1,8 @@
+export default function Portfolio (){
+    return(
+        <div>
+            <h1>Portfolio</h1>
+            
+        </div>
+    )
+}
