+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import styles from '../styles/Home.module.scss';
 import HomePage from '../components/home_page';
 import About from '../components/about';
-import Area from '../components/experiance';
+import Area from '../components/expertise';
 
 
 export default function Home() {
